+++ src/main/kotlin/com/github/ixmoyren/typilot/Constants.kt
@@ -1,5 +1,3 @@
 package com.github.ixmoyren.typilot
 
 internal const val TYPILOT_NOTIFICATION_GROUP_ID = "Typilot"
-
-internal const val TYPST_LANGUAGE_SERVER_ID = "tinymistServer"