+++ src/main/kotlin/com/github/ixmoyren/typilot/actions/TinymistExportPdfAction.kt
@@ -1,8 +1,7 @@
 package com.github.ixmoyren.typilot.actions
 
-import com.github.ixmoyren.typilot.TYPST_LANGUAGE_SERVER_ID
 import com.github.ixmoyren.typilot.language.TypstFileType
-import com.github.ixmoyren.typilot.lsp.TinymistExportPdfHandler
+import com.github.ixmoyren.typilot.lsp.TinymistCommands
 import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
@@ -19,6 +18,6 @@ class TinymistExportPdfAction : AnAction() {
     override fun actionPerformed(event: AnActionEvent) {
         val project = event.project ?: return
         val file = event.getData(CommonDataKeys.VIRTUAL_FILE)?.takeIf { it.fileType == TypstFileType } ?: return
-        TinymistExportPdfHandler.perform(project, file, null, TYPST_LANGUAGE_SERVER_ID)
+        TinymistCommands.exportPdf(project, file)
     }
 }